Split BasicOp into two classes to aloow for op without value
